### 2025-08-25 f5b0926
Add CRE Model
