### 2025-08-26 #8867 @editorialbot
[REVIEW]: pprof: An R Package for Provider Profiling
**Submitting author:** @EmiiilyLiu (<a href="http://orcid.org/0009-0005-5397-6810">Xiaohan Liu</a>)
**Repository:** https://github.com/UM-KevinHe/pprof
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 1.0.2
