### 2025-12-23 #8541 @soodoku
I have done a pass and opened issue. See here: https://github.com/UM-KevinHe/pprof/issues
