### 2026-01-14 #8867 @skanwal
Hello [@eurunuela](https://github.com/eurunuela) [@MadhviKannan](https://github.com/MadhviKannan) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
