+++ src/logis_firth.cpp
@@ -1,0 +1,473 @@
+//#define ARMA_NO_DEBUG
+#define STRICT_R_HEADERS // needed on Windows, not on macOS
+#include <RcppParallel.h>
+// [[Rcpp::depends(RcppParallel)]]
+#define ARMA_DONT_USE_OPENMP
+#include <RcppArmadillo.h>
+// [[Rcpp::depends(RcppArmadillo)]]
+#include <cmath>
+#include "header.h"
+#include "myomp.h"
+#include <iostream>
+#include <chrono>
+
+// [[Rcpp::plugins(cpp11)]]
+// [[Rcpp::plugins(openmp)]]
+using namespace RcppParallel;
+using namespace Rcpp;
+using namespace std;
+using namespace arma;
+
+//` @importFrom RcppParallel RcppParallelLibs
+
+
+// [[Rcpp::export]]
+List logis_LR_prov(arma::vec &Y, arma::mat &Z, arma::vec &n_prov, arma::vec gamma, arma::vec beta,
+                   int n_obs, int m, const int &threads, double tol = 1e-5, int max_iter = 100,
+                   double bound = 10.0, bool message = true, //bool backtrack = false,
+                   const std::string stop = "beta") {
+
+  arma::ivec indices(m + 1);
+  int ind = 0;
+  for (int i = 0; i < m; i++) {
+    indices(i) = ind;
+    ind += n_prov(i);
+  }
+  indices(m) = ind;
+
+  arma::vec gamma_obs = rep(gamma, n_prov);
+  double loglkd_init = Loglkd(Y, Z*beta, gamma_obs);
+
+  int chunk_id;
+
+  int beta_size = Z.n_cols;
+  arma::vec score_beta(beta_size);
+  arma::vec score_gamma(m);
+  arma::vec info_gamma_inv(m);
+  arma::mat J(beta_size, m);
+  arma::mat J2(beta_size, m);  //I^*(\beta)^{-1} * J; where "I^*(\beta)^{-1}" is "schur_inv".
+  arma::mat schur(beta_size, beta_size);  //I^*(\beta)
+  arma::mat schur_inv(beta_size, beta_size);   //I^*(\beta)^{-1}
+  arma::vec p(n_obs);
+  arma::vec Yp(n_obs);
+  arma::vec pq(n_obs);
+  arma::vec H(beta_size);
+
+  arma::vec old_gamma = gamma;
+
+
+  omp_set_num_threads(threads);
+  int iter = 0;
+  double crit = 100.0, d_loglkd = 0.0, old_loglkd = loglkd_init, loglkd = loglkd_init;
+
+  #pragma omp parallel shared(indices, iter, crit)
+  {
+    arma::mat local_schur(beta_size, beta_size, arma::fill::zeros);
+    arma::vec local_H(beta_size, arma::fill::zeros);
+    while (iter < max_iter && crit > tol) {
+        #pragma omp single
+        {
+            chunk_id = 0;
+            schur.zeros();
+        }
+        #pragma omp barrier
+
+        while (true){
+            int my_chunk;
+            #pragma omp critical (get_chunk_id_1)
+            {
+                if (chunk_id < m) {
+                    my_chunk = chunk_id;
+                    chunk_id++;
+                } else {
+                    my_chunk = -1;
+                }
+            }
+            if (my_chunk == -1) {
+                break;
+            }
+
+            // Get the chunk of Z and Y
+            // these "subviews" provide a window into the original matrices without copying data.
+            const arma::mat &TDS_Z = Z.rows(indices(my_chunk), indices(my_chunk + 1) - 1);
[…]
