### 2026-02-02 #8867 @jungtaekkim
Hi @skanwal,

I am now reviewing a paper for JOSS.  Since a thorough review takes significant time, I’m not able to take on another paper right now.
