### 2026-02-02 #8867 @cmaimone
I took a look quickly at the package and paper.  While I was able to install the package, when I went to load it, there was something missing from the dependencies, so I needed to install another package (solved by installing globals):

```
> library(pprof)
Error: package or namespace load failed for ‘pprof’ in loadNamespace(j <- i[[1L]], c(lib.loc, .libPaths()), versionCheck = vI[[j]]):
 there is no package called ‘globals’
```

From a quick skim through the package and paper, I'm confused about what the package is for.  It's framed as being for estimation of healthcare provider effects, but it appears to be reimplementing standard regression models to be more computationally efficient for a large number of fixed effects, which is a general problem (among potentially additional contributions).  This general application of the package is backed up by the example given in the paper, which isn't about healthcare providers or using healthcare data.

If it is implementing more computationally efficient methods, then there should be performance profiling included, as well as more extensive testing of the correctness of the methods.  A scan of tests in the package did not appear to be testing the correctness of the implemented algorithms.

I hope these quick comments can help move the package review forward, but I unfortunately don't have time to be a formal reviewer at the moment given these concerns.
