+++ man/logis_firth.Rd
@@ -88,6 +88,10 @@ The function accepts three different input formats:
 a formula and dataset, where the formula is of the form \code{response ~ covariates + id(provider)}, with \code{provider} representing the provider identifier;
 a dataset along with the column names of the response, covariates, and provider identifier;
 or the binary outcome vector \eqn{\boldsymbol{Y}}, the covariate matrix or data frame \eqn{\mathbf{Z}}, and the provider identifier vector.
+
+This function utilizes OpenMP for parallel processing. For macOS, to enable multi-threading,
+users may need to install the OpenMP library (e.g., brew install libomp) or use a supported compiler such as GCC.
+If OpenMP is not detected during installation, the function will transparently fall back to single-threaded execution.
 }
 \examples{
 data(ExampleDataBinary)
