+++ NAMESPACE
@@ -45,6 +45,7 @@ importFrom(Matrix,bdiag)
 importFrom(Rcpp,evalCpp)
 importFrom(RcppParallel,RcppParallelLibs)
 importFrom(caret,nearZeroVar)
+importFrom(dplyr,across)
 importFrom(dplyr,arrange)
 importFrom(dplyr,bind_rows)
 importFrom(dplyr,cross_join)
@@ -123,4 +124,5 @@ importFrom(stats,setNames)
 importFrom(stats,terms)
 importFrom(stats,uniroot)
 importFrom(tibble,tibble)
+importFrom(tidyselect,all_of)
 useDynLib(pprof, .registration = TRUE)
