+++ R/logis_firth.R
@@ -81,7 +81,9 @@
 #' @importFrom stats complete.cases terms model.matrix reformulate median
 #'
 #' @references
-#' He K,
+#' Firth, D. (1993) Bias reduction of maximum likelihood estimates.
+#' \emph{Biometrika}, \strong{80(1)}: 27-38.
+#' \cr
 #'
 #' @export
 #'
