+++ src/Fixed_effect.cpp
@@ -120,252 +120,6 @@ arma::vec computeDirectExp(const arma::vec& est, const arma::vec& Z_beta, const 
     return results;
 }
 
-/*
-// [[Rcpp::export]]
-List logis_firth_prov(arma::vec &Y, arma::mat &Z, arma::vec &n_prov, arma::vec gamma, arma::vec beta,
-                      int n_obs, int m, const int &threads, double tol = 1e-5, int max_iter = 100,
-                      double bound = 10.0, bool message = true, bool firth = true, bool gamma_tol = false) {
-  if (message == true) {
-    cout << "Implementing firth-corrected fixed provider effects model (Rcpp) ..." << endl;
-  }
-
-  // make "provider" as the basic unit of parallelization (so ideally maximum number of threads = number of providers)
-  arma::ivec indices(m + 1);
-  int ind = 0;
-  for (int i = 0; i < m; i++) {
-    indices(i) = ind;
-    ind += n_prov(i);
-  }
-  indices(m) = ind;
-
-  int chunk_id; //each thread will process a chunk of data
-
-  int beta_size = Z.n_cols;
-  arma::vec score_beta(beta_size);
-  arma::vec score_gamma(m);
-  arma::vec info_gamma_inv(m);
-  arma::mat J(beta_size, m);
-  arma::mat J2(beta_size, m);  //I^*(\beta)^{-1} * J; where "I^*(\beta)^{-1}" is "schur_inv".
-  arma::mat schur(beta_size, beta_size);  //I^*(\beta)
-  arma::mat schur_inv(beta_size, beta_size);   //I^*(\beta)^{-1}
-  arma::vec p(n_obs);
-  arma::vec Yp(n_obs);
-  arma::vec pq(n_obs);
-  arma::vec H(beta_size);
-
-  arma::vec old_gamma = gamma;
-
-
-  omp_set_num_threads(threads);
-  int iter = 0;
-  double crit = 100.0;
-  double crit2 = 100.0;
-
-  #pragma omp parallel shared(indices, iter, crit)
-  {
-    arma::mat local_schur(beta_size, beta_size, arma::fill::zeros);
-    arma::vec local_H(beta_size, arma::fill::zeros);
-    while (iter < max_iter && crit > tol) {
-      #pragma omp single
-      {
-        chunk_id = 0;
-        schur.zeros();
-      }
-      #pragma omp barrier
-
-      while (true){
-        int my_chunk;
-        #pragma omp critical (get_chunk_id_1)
-        {
-          if (chunk_id < m) {
-            my_chunk = chunk_id;
-            chunk_id++;
-          } else {
-            my_chunk = -1;
-          }
-        }
-        if (my_chunk == -1) {
-          break;
-        }
-
-        // Get the chunk of Z and Y
-        // these "subviews" provide a window into the original matrices without copying data.
-        const arma::mat &TDS_Z = Z.rows(indices(my_chunk), indices(my_chunk + 1) - 1);
-        const arma::vec &TDS_Y = Y.subvec(indices(my_chunk), indices(my_chunk + 1) - 1);
-
-        arma::vec TDS_gamma_obs = rep(gamma(my_chunk), n_prov(my_chunk)); // get the chunk of gamma
-        arma::vec TDS_Z_beta = TDS_Z * beta;
-        arma::vec TDS_p = 1 / (1 + exp(-TDS_gamma_obs-TDS_Z_beta));
-        arma::vec TDS_Yp = TDS_Y - TDS_p;
-        arma::vec TDS_pq = TDS_p % (1 - TDS_p);
-        if (arma::any(TDS_pq == 0)) {
-          TDS_pq.replace(0, 1e-10);
-        }
-
[…]
