+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: pprof
 Type: Package
 Title: Modeling, Standardization and Testing for Provider Profiling
-Version: 1.0.2
-Date: 2025-06-20
+Version: 1.0.3
+Date: 2026-02-08
 Authors@R: c(
    person("Xiaohan", "Liu", email = "xhliuu@umich.edu", role = c("aut", "cre")),
    person("Lingfeng", "Luo", email = "lfluo@umich.edu", role = "aut"),
@@ -35,7 +35,8 @@ Imports:
     scales,
     tibble,
     rlang,
-    tidyselect
+    tidyselect,
+    globals
 LinkingTo: Rcpp, RcppArmadillo, RcppParallel
 Depends:
     R (>= 4.1.0)

[…]
