### 2026-02-27 #8867 @skanwal
👋 [@Ruth-Moraa](https://github.com/Ruth-Moraa)ME - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
