### 2026-03-22 #8867 @ammawla
> Hello [@ammawla](https://github.com/ammawla) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @skanwal, I'd be happy to.  What is the expected turn around time?
