### 2026-04-21 #8867 @ammawla
@skanwal I completed the review, above: https://github.com/openjournals/joss-reviews/issues/8867#issuecomment-4150573307.

I created a tagged issue, also above: https://github.com/UM-KevinHe/pprof/issues/13. I have not received any updates regarding those issues being addressed.
